requirements.txt
Django==3.2.8
django-dotenv==1.4.2
django-filter==21.1
django-ses==3.3.0
djangorestframework==3.12.4
python-dotenv==0.20.0
django-cors-headers==4.4.0